LayoutTests/fast/viewport/viewport-47.html

<head>
    <title>Check that width is increased to fill the visual viewport.</title>
    <meta name="viewport" content="width=10, height=3000, initial-scale=1.0">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
