Safari-611.1.10/LayoutTests/fast/viewport/viewport-47-expected.txt

ALERT: viewport size 320x3000 scale 1 with limits [1, 5] and userScalable true
